+++ CopilotKit/packages/runtime/src/lib/runtime/remote-action-constructors.ts
@@ -60,10 +60,12 @@ export function constructLGCRemoteAction({
       });
 
       let state = {};
+      let configurable = {};
       if (agentStates) {
-        const jsonState = agentStates.find((state) => state.agentName === name)?.state;
+        const jsonState = agentStates.find((state) => state.agentName === name);
         if (jsonState) {
-          state = JSON.parse(jsonState);
+          state = JSON.parse(jsonState.state);
+          configurable = JSON.parse(jsonState.configurable);
         }
       }
 
@@ -77,6 +79,7 @@ export function constructLGCRemoteAction({
           nodeName,
           messages: [...messages, ...additionalMessages],
           state,
+          configurable,
           properties: graphqlContext.properties,
           actions: actionInputsWithoutAgents.map((action) => ({
             name: action.name,
@@ -206,10 +209,12 @@ export function constructRemoteActions({
           });
 
           let state = {};
+          let configurable = {};
           if (agentStates) {
-            const jsonState = agentStates.find((state) => state.agentName === name)?.state;
+            const jsonState = agentStates.find((state) => state.agentName === name);
             if (jsonState) {
-              state = JSON.parse(jsonState);
+              state = JSON.parse(jsonState.state);
+              configurable = JSON.parse(jsonState.configurable);
             }
           }
 
@@ -224,6 +229,7 @@ export function constructRemoteActions({
                 nodeName,
                 messages: [...messages, ...additionalMessages],
                 state,
+                configurable,
                 properties: graphqlContext.properties,
                 actions: actionInputsWithoutAgents.map((action) => ({
                   name: action.name,