+++ examples/coagents-qa-native/ui/app/Mailer.tsx
@@ -73,6 +73,14 @@ export function Mailer() {
     render: ({ event, resolve }) => (
       <InterruptForm event={event} resolve={resolve} />
     ),
+    enabled: ({ eventValue, agentMetadata }) => {
+      return (
+        eventValue ===
+          "Please provide a sender name which will appear in the email" &&
+        agentMetadata.agentName === "email_agent" &&
+        agentMetadata.nodeName === "email_node"
+      );
+    },
   });
 
   return (