+++ examples/langgraph-tutorial-customer-support/ui/src/app/page.tsx
@@ -1,0 +1,21 @@
+"use client";
+import { useCopilotAction } from "@copilotkit/react-core";
+import { CopilotPopup } from "@copilotkit/react-ui";
+import { useState } from "react";
+
+export default function YourApp() {
+  return (
+    <>
+      <CopilotPopup
+        instructions={
+          "You are assisting the user as best as you can. Answer in the best way possible given the data you have."
+        }
+        defaultOpen={true}
+        labels={{
+          title: "Popup Assistant",
+          initial: "Need any help?",
+        }}
+      />
+    </>
+  );
+}