+++ infra/lib/utils.ts
@@ -20,27 +20,21 @@ export function createAgentProjectStack({
   app,
   project,
   description,
-  dependencies,
 }: {
   app: App;
   project: string;
   description: string;
-  dependencies: "Remote" | "Local";
 }): {
   selfHostedAgent: PreviewProjectStack;
   lgcAgentPython: PreviewProjectStack;
   lgcAgentJS: PreviewProjectStack;
 } {
   const cdkStackName =
-    toCdkStackName(project) + "Agent" + dependencies + "Deps";
-  const dockerfile =
-    dependencies === "Remote"
-      ? `examples/Dockerfile.agent-remote-deps`
-      : `examples/Dockerfile.agent-local-deps`;
+    toCdkStackName(project) + "AgentLocalDeps";
   const GITHUB_ACTIONS_RUN_ID = requireEnv("GITHUB_ACTIONS_RUN_ID");
 
   const outputs: Record<string, string> = {
-    Dependencies: dependencies,
+    Dependencies: "Local",
   };
 
   if (process.env.GITHUB_PR_NUMBER) {
@@ -50,8 +44,6 @@ export function createAgentProjectStack({
   const selfHostedAgent = new PreviewProjectStack(app, cdkStackName, {
     projectName: project,
     projectDescription: description,
-    demoDir: `examples/${project}/agent`,
-    overrideDockerfile: dockerfile,
     environmentVariablesFromSecrets: [
       "OPENAI_API_KEY",
       "ANTHROPIC_API_KEY",
@@ -63,9 +55,7 @@ export function createAgentProjectStack({
     env: {
       account: process.env.CDK_DEFAULT_ACCOUNT,
     },
-    imageTag: `${project}-agent-python-${
-      dependencies === "Remote" ? "remote-deps" : "local-deps"
-    }-${GITHUB_ACTIONS_RUN_ID}`,
+    imageTag: `${project}-agent-python-local-deps-${GITHUB_ACTIONS_RUN_ID}`,
     outputs: {
       ...outputs,
       LangGraphCloud: "false",
@@ -76,8 +66,6 @@ export function createAgentProjectStack({
   const lgcAgentPython = new PreviewProjectStack(app, `${cdkStackName}LGCPython`, {
     projectName: project,
     projectDescription: `${description} - LangGraph Cloud Python`,
-    demoDir: `examples/${project}/agent`,
-    overrideDockerfile: dockerfile,
     environmentVariablesFromSecrets: [
       "OPENAI_API_KEY",
       "ANTHROPIC_API_KEY",
@@ -90,9 +78,7 @@ export function createAgentProjectStack({
     env: {
       account: process.env.CDK_DEFAULT_ACCOUNT,
     },
-    imageTag: `${project}-agent-python-${
-      dependencies === "Remote" ? "remote-deps" : "local-deps"
-    }-${GITHUB_ACTIONS_RUN_ID}`,
+    imageTag: `${project}-agent-python-local-deps-${GITHUB_ACTIONS_RUN_ID}`,
     entrypoint: ["/bin/sh", "-c"],
     cmd: ["langgraph dev --no-browser --port=8000 --config=langgraph.json --host=0.0.0.0"],
     outputs: {
@@ -105,8 +91,6 @@ export function createAgentProjectStack({
   const lgcAgentJS = new PreviewProjectStack(app, `${cdkStackName}LGCJS`, {
     projectName: project,
     projectDescription: `${description} - LangGraph Cloud JS`,
-    demoDir: `examples/${project}/agent`,
-    overrideDockerfile: dockerfile,
     environmentVariablesFromSecrets: [
       "OPENAI_API_KEY",
       "ANTHROPIC_API_KEY",
@@ -114,16 +98,19 @@ export function createAgentProjectStack({
       "TAVILY_API_KEY",
       "LANGSMITH_API_KEY",
     ],
+    environmentVariables: {
+      HOME: "/tmp",
+      PNPM_HOME: "/tmp/pnpm",
+      COREPACK_HOME: "/tmp/corepack",
+    },
     port: "8000",
     includeInPRComment: false,
     env: {
       account: process.env.CDK_DEFAULT_ACCOUNT,
     },
-    imageTag: `${project}-agent-js-${
-      dependencies === "Remote" ? "remote-deps" : "local-deps"
-    }-${GITHUB_ACTIONS_RUN_ID}`,
-    entrypoint: ["/bin/sh", "-c"],
-    cmd: ["pnpx @langchain/langgraph-cli dev --config=langgraph.json --no-browser --port 8000 --host 0.0.0.0"],
+    imageTag: `${project}-agent-js-local-deps-${GITHUB_ACTIONS_RUN_ID}`,
+    // entrypoint: ["/bin/sh", "-c"],
+    // cmd: ["pnpx @langchain/langgraph-cli dev --config=langgraph.json --no-browser --port 8000 --host 0.0.0.0"],
     outputs: {
       ...outputs,
       LangGraphCloud: "false",
@@ -138,7 +125,6 @@ export function createUIProjectStack({
   app,
   project,
   description,
-  dependencies,
   selfHostedAgentProject,
   lgcAgentProjectPython,
   lgcAgentProjectJS,
@@ -149,7 +135,6 @@ export function createUIProjectStack({
   app: App;
   project: string;
   description: string;
-  dependencies: "Remote" | "Local";
   selfHostedAgentProject: PreviewProjectStack;
   lgcAgentProjectPython: PreviewProjectStack;
   lgcAgentProjectJS: PreviewProjectStack;
@@ -157,16 +142,12 @@ export function createUIProjectStack({
   environmentVariablesFromSecrets?: string[];
   customOutputs?: Record<string, string>;
 }) {
-  const cdkStackName = toCdkStackName(project) + "UI" + dependencies + "Deps";
-  const dockerfile =
-    dependencies === "Remote"
-      ? `examples/Dockerfile.ui-remote-deps`
-      : `examples/Dockerfile.ui-local-deps`;
+  const cdkStackName = toCdkStackName(project) + "UILocalDeps";
   const GITHUB_ACTIONS_RUN_ID = requireEnv("GITHUB_ACTIONS_RUN_ID");
 
   const outputs: Record<string, string> = {
-    Dependencies: dependencies,
-    EndToEndProjectKey: `${project}-ui-deps-${dependencies.toLocaleLowerCase()}`,
+    Dependencies: "Local",
+    EndToEndProjectKey: `${project}-ui-deps-local`,
     LgcPythonDeploymentUrl: `${lgcAgentProjectPython.fnUrl}`,
     LgcJSDeploymentUrl: `${lgcAgentProjectJS.fnUrl}`,
   };
@@ -182,8 +163,6 @@ export function createUIProjectStack({
   return new PreviewProjectStack(app, cdkStackName, {
     projectName: project,
     projectDescription: `${description}`,
-    demoDir: `examples/${project}/ui`,
-    overrideDockerfile: dockerfile,
     environmentVariablesFromSecrets: [
       "OPENAI_API_KEY",
       "LANGSMITH_API_KEY",
@@ -213,12 +192,11 @@ export function createNextOpenAIProjectStack({
 }: {
   app: App;
   description: string;
-  variant: "self-hosted" | "against-cloud-prod" | "against-cloud-staging";
+  variant: "self-hosted";
   environmentVariables?: Record<string, string>;
   environmentVariablesFromSecrets?: string[];
 }) {
   const cdkStackName = toCdkStackName(`next-openai-${variant}`);
-  const dockerfile = `CopilotKit/examples/next-openai/Dockerfile`;
   const GITHUB_ACTIONS_RUN_ID = requireEnv("GITHUB_ACTIONS_RUN_ID");
 
   const outputs: Record<string, string> = {
@@ -234,8 +212,6 @@ export function createNextOpenAIProjectStack({
   return new PreviewProjectStack(app, cdkStackName, {
     projectName: `next-openai`,
     projectDescription: `${description}`,
-    demoDir: `CopilotKit/examples/next-openai`,
-    overrideDockerfile: dockerfile,
     environmentVariablesFromSecrets: [
       "OPENAI_API_KEY",
       "ANTHROPIC_API_KEY",