+++ sdk-python/copilotkit/integrations/fastapi.py
@@ -192,6 +192,7 @@ async def handler_v1(
     if method == 'POST' and path == 'agents/execute':
         thread_id = body.get("threadId")
         node_name = body.get("nodeName")
+        configurable = body.get("configurable")
 
         name = body_get_or_raise(body, "name")
         state = body_get_or_raise(body, "state")
@@ -206,6 +207,7 @@ async def handler_v1(
             node_name=node_name,
             name=name,
             state=state,
+            configurable=configurable,
             messages=messages,
             actions=actions,
             meta_events=meta_events,
@@ -270,6 +272,7 @@ def handle_execute_agent( # pylint: disable=too-many-arguments
         thread_id: str,
         name: str,
         state: dict,
+        configurable: Optional[dict] = None,
         messages: List[Message],
         actions: List[ActionDict],
         node_name: str,
@@ -283,6 +286,7 @@ def handle_execute_agent( # pylint: disable=too-many-arguments
             name=name,
             node_name=node_name,
             state=state,
+            configurable=configurable,
             messages=messages,
             actions=actions,
             meta_events=meta_events,